Extract Modal into a separate page, Product-Details
Remove unnecessary code/dependencies from the modal
