+++ package.json
@@ -20,7 +20,6 @@
     "react-emotion": "^9.1.3",
     "react-helmet": "^5.2.0",
     "react-icons": "^2.2.7",
-    "react-modal": "^3.4.5",
     "react-onclickoutside": "^6.7.1",
     "react-router-dom": "^4.3.1",
     "recompose": "^0.27.1",
