+++ src/components/shared/Layout.js
@@ -175,7 +175,7 @@ export default class Layout extends React.Component {
             <Header />
             {!this.state.user.profile.name && <CTA />}
             <Main>{this.props.children}</Main>
-            <Footer displayAbout={this.props.location.pathname === '/'} />
+            <Footer displayAbout={this.props.location.pathname === '/' || this.props.location.pathname === '/product-details'} />
           </StoreContext.Provider>
         </UserContext.Provider>
       </>
